chatbot-api/requirements.txt
algoliasearch>=4,<5
python-dotenv
openai
agno
memory_profiler
nltk
pandas
fastapi
uvicorn
httpx
qdrant-client[fastembed]
tiktoken
fastembed
mysql-connector-python>=8.3,<9
